Fix the unauthorized player deletion

+++ src/main/java/ro/mg/chessserver/controller/PlayerController.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -70,8 +71,12 @@ public class PlayerController {
             return ResponseEntity.status(HttpStatus.OK).body(updated);
     }
 
-    @DeleteMapping
-    public ResponseEntity<Void> delete(@RequestParam long id) {
+    @DeleteMapping()
+    public ResponseEntity<Void> delete(@RequestParam long id, Authentication auth) {
+        long loggedInUserId = Long.parseLong(auth.getName());
+        if (loggedInUserId != id)
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+
         playerService.deletePlayer(id);
         return ResponseEntity.ok().build();
     }
